+++ server/api/footer-link/delete-by-id/[id].get.ts
@@ -2,13 +2,26 @@ import { defineEventHandler } from "#imports";
 
 export default defineEventHandler(async (event) => {
   try {
-    await event.context.prisma.footerLink.delete({
-      where: {
-        id: event?.context?.params?.id,
-      },
-    });
+    if (event?.context?.params?.id) {
+      const result = await event.context.prisma.footerLink.delete({
+        where: {
+          id: event?.context?.params?.id,
+        },
+      });
 
-    return "Success";
+      return {
+        statusCode: 200,
+        statusMessage: "Success",
+        table: "footer-link",
+        method: "delete",
+        objectResult: result,
+      };
+    } else {
+      throw createError({
+        statusCode: 400,
+        statusMessage: "The ID file is empty in request",
+      });
+    }
   } catch (error) {
     console.log(error);
     return error;
